Added stubs and proper logic
