+++ requirements.test.txt
@@ -5,3 +5,4 @@ mypy-extensions==0.4.1
 pytest==4.3.1
 pytest-cov==2.8.1
 codecov==2.0.15
+hvac==0.10.9
